[REFACTOR] Move character data to own file
